+++ uss/drivers/ch34x/ch34x.hpp
@@ -94,13 +94,13 @@ public:
             throw error::InvalidDeviceConfigException("(1)stop_bits",
                                                       (int)device.stop_bits);
 
-        if ((int)device.data_bits >= sizeof(DataBitsConverter))
+        if ((uint32_t)device.data_bits >= sizeof(DataBitsConverter))
             throw error::InvalidDeviceConfigException("data_bits",
                                                       (int)device.data_bits);
-        if ((int)device.stop_bits >= sizeof(StopBitsConverter))
+        if ((uint32_t)device.stop_bits >= sizeof(StopBitsConverter))
             throw error::InvalidDeviceConfigException("(2)stop_bits",
                                                       (int)device.stop_bits);
-        if ((int)device.parity >= sizeof(ParityConverter))
+        if ((uint32_t)device.parity >= sizeof(ParityConverter))
             throw error::InvalidDeviceConfigException("parity",
                                                       (int)device.parity);
 
@@ -218,7 +218,7 @@ public:
 
         // For each configuration.. (with the amount of them found in the device
         // descriptor)
-        for (int ic = 0; ic < device_descriptor.bNumConfigurations; ic++) {
+        for (uint8_t ic = 0; ic < device_descriptor.bNumConfigurations; ic++) {
             // Get the configuration descriptor
             libusb_get_config_descriptor(device.GetUsbDevice(), ic,
                                          &config_descriptor);
